UnaryMinusAndNegate work.
