+++ src/engine/sparqlExpressions/NaryExpression.cpp
@@ -3,61 +3,10 @@
 //  Author: Johannes Kalmbach <kalmbacj@cs.uni-freiburg.de>
 
 #include "engine/sparqlExpressions/NaryExpression.h"
+#include "engine/sparqlExpressions/NaryExpressionImpl.h"
 
 namespace sparqlExpression {
 namespace detail {
-
-// _____________________________________________________________________________
-template <typename Op>
-requires(isOperation<Op>)
-NaryExpression<Op>::NaryExpression(Children&& children)
-    : _children{std::move(children)} {}
-
-// _____________________________________________________________________________
-
-template <typename NaryOperation>
-requires(isOperation<NaryOperation>)
-ExpressionResult NaryExpression<NaryOperation>::evaluate(
-    EvaluationContext* context) const {
-  auto resultsOfChildren = ad_utility::applyFunctionToEachElementOfTuple(
-      [context](const auto& child) { return child->evaluate(context); },
-      _children);
-
-  // A function that only takes several `ExpressionResult`s,
-  // and evaluates the expression.
-  auto evaluateOnChildrenResults =
-      std::bind_front(ad_utility::visitWithVariantsAndParameters,
-                      evaluateOnChildrenOperands, NaryOperation{}, context);
-
-  return std::apply(evaluateOnChildrenResults, std::move(resultsOfChildren));
-}
-
-// _____________________________________________________________________________
-template <typename Op>
-requires(isOperation<Op>)
-std::span<SparqlExpression::Ptr> NaryExpression<Op>::children() {
-  return {_children.data(), _children.size()};
-}
-
-// __________________________________________________________________________
-template <typename Op>
-requires(isOperation<Op>) [[nodiscard]] string NaryExpression<Op>::getCacheKey(
-    const VariableToColumnMap& varColMap) const {
-  string key = typeid(*this).name();
-  for (const auto& child : _children) {
-    key += child->getCacheKey(varColMap);
-  }
-  return key;
-}
-
-#define NARY_EXPRESSION(Name, N, X, ...)                                     \
-  class Name : public NaryExpression<detail::Operation<N, X, __VA_ARGS__>> { \
-    using Base = NaryExpression<Operation<N, X, __VA_ARGS__>>;               \
-    using Base::Base;                                                        \
-  };
-
-#define INSTANTIATE_NARY(N, X, ...) \
-  template class NaryExpression<Operation<N, X, __VA_ARGS__>>
 
 INSTANTIATE_NARY(2, FV<decltype(orLambda), EffectiveBooleanValueGetter>,
                  SET<SetOfIntervals::Union>);
@@ -78,11 +27,6 @@ INSTANTIATE_NARY(2, FV<decltype(add), NumericValueGetter>);
 
 INSTANTIATE_NARY(2, FV<decltype(subtract), NumericValueGetter>);
 
-INSTANTIATE_NARY(1, FV<decltype(abs), NumericValueGetter>);
-INSTANTIATE_NARY(1, FV<decltype(round), NumericValueGetter>);
-INSTANTIATE_NARY(1, FV<decltype(ceil), NumericValueGetter>);
-INSTANTIATE_NARY(1, FV<decltype(floor), NumericValueGetter>);
-
 INSTANTIATE_NARY(1,
                  FV<NumericIdWrapper<decltype(ad_utility::wktLongitude), true>,
                     StringValueGetter>);
@@ -98,8 +42,6 @@ INSTANTIATE_NARY(1, FV<decltype(extractDay), DateValueGetter>);
 INSTANTIATE_NARY(1, FV<std::identity, StringValueGetter>);
 INSTANTIATE_NARY(1, FV<decltype(strlen), StringValueGetter>);
 
-NARY_EXPRESSION(AbsExpressionInternal, 1,
-                FV<decltype(abs), NumericValueGetter>);
 }  // namespace detail
 
 SparqlExpression::Ptr makeAddExpression(SparqlExpression::Ptr child1,
@@ -120,8 +62,13 @@ SparqlExpression::Ptr makeOrExpression(SparqlExpression::Ptr child1,
 SparqlExpression::Ptr makeSubtractExpression(SparqlExpression::Ptr child1,
                                              SparqlExpression::Ptr child2);
 
-SparqlExpression::Ptr makeUnaryMinusExpression(SparqlExpression::Ptr child);
-SparqlExpression::Ptr makeUnaryNegateExpression(SparqlExpression::Ptr child);
+SparqlExpression::Ptr makeUnaryMinusExpression(SparqlExpression::Ptr child) {
+  return std::make_unique<UnaryMinusExpression>(std::move(child));
+}
+
+SparqlExpression::Ptr makeUnaryNegateExpression(SparqlExpression::Ptr child) {
+  return std::make_unique<UnaryNegateExpression>(std::move(child));
+}
 
 SparqlExpression::Ptr makeRoundExpression(SparqlExpression::Ptr child);
 SparqlExpression::Ptr makeAbsExpression(SparqlExpression::Ptr child);
@@ -139,16 +86,4 @@ SparqlExpression::Ptr makeYearExpression(SparqlExpression::Ptr child);
 SparqlExpression::Ptr makeStrExpression(SparqlExpression::Ptr child);
 SparqlExpression::Ptr makeStrlenExpression(SparqlExpression::Ptr child);
 
-SparqlExpression::Ptr makeRoundExpression(SparqlExpression::Ptr child) {
-  return std::make_unique<RoundExpression>(std::move(child));
-}
-SparqlExpression::Ptr makeAbsExpression(SparqlExpression::Ptr child) {
-  return std::make_unique<detail::AbsExpressionInternal>(std::move(child));
-}
-SparqlExpression::Ptr makeCeilExpression(SparqlExpression::Ptr child) {
-  return std::make_unique<CeilExpression>(std::move(child));
-}
-SparqlExpression::Ptr makeFloorExpression(SparqlExpression::Ptr child) {
-  return std::make_unique<FloorExpression>(std::move(child));
-}
 }  // namespace sparqlExpression
