+++ src/engine/sparqlExpressions/NumericExpressions.cpp
@@ -1,0 +1,70 @@
+//  Copyright 2023, University of Freiburg,
+//                  Chair of Algorithms and Data Structures.
+//  Author: Johannes Kalmbach <kalmbacj@cs.uni-freiburg.de>
+
+#include "engine/sparqlExpressions/NaryExpressionImpl.h"
+
+namespace sparqlExpression {
+namespace detail {
+// Abs
+inline const auto absImpl = []<typename T>(T num) { return std::abs(num); };
+inline const auto abs = makeNumericExpression<decltype(absImpl)>();
+NARY_EXPRESSION(AbsExpression, 1,
+                FV<decltype(abs), NumericValueGetter>);
+
+// Additional numeric expressions
+// Round
+inline const auto roundImpl = []<typename T>(T num) {
+  if constexpr (std::is_floating_point_v<T>) {
+    auto res = std::round(num);
+    // In SPARQL negative numbers are rounded towards zero if they lie exactly
+    // between two integers.
+    if (num < 0 && std::abs(res - num) == 0.5) {
+      res += 1;
+    }
+    return res;
+  } else {
+    return num;
+  }
+};
+
+inline const auto round = makeNumericExpression<decltype(roundImpl)>();
+NARY_EXPRESSION(RoundExpression, 1, FV<decltype(round), NumericValueGetter>);
+
+// Ceil
+inline const auto ceilImpl = []<typename T>(T num) {
+  if constexpr (std::is_floating_point_v<T>) {
+    return std::ceil(num);
+  } else {
+    return num;
+  }
+};
+inline const auto ceil = makeNumericExpression<decltype(ceilImpl)>();
+NARY_EXPRESSION(CeilExpression, 1, FV<decltype(ceil), NumericValueGetter>);
+
+// Floor
+inline const auto floorImpl = []<typename T>(T num) {
+  if constexpr (std::is_floating_point_v<T>) {
+    return std::floor(num);
+  } else {
+    return num;
+  }
+};
+inline const auto floor = makeNumericExpression<decltype(floorImpl)>();
+using FloorExpression = NARY<1, FV<decltype(floor), NumericValueGetter>>;
+}
+
+using namespace detail;
+SparqlExpression::Ptr makeRoundExpression(SparqlExpression::Ptr child) {
+  return std::make_unique<RoundExpression>(std::move(child));
+}
+SparqlExpression::Ptr makeAbsExpression(SparqlExpression::Ptr child) {
+  return std::make_unique<AbsExpression>(std::move(child));
+}
+SparqlExpression::Ptr makeCeilExpression(SparqlExpression::Ptr child) {
+  return std::make_unique<CeilExpression>(std::move(child));
+}
+SparqlExpression::Ptr makeFloorExpression(SparqlExpression::Ptr child) {
+  return std::make_unique<FloorExpression>(std::move(child));
+}
+}
